+++ docs/release-notes.md
@@ -9,8 +9,16 @@
 
 ## Upcoming release
 * For players:
+  * Updated for Stardew Valley 1.5, including split-screen support.
   * When the installer is run from within a game folder, it now installs SMAPI to that folder. That simplifies installation if you have multiple copies of the game or it can't otherwise auto-detect the game path.
   * Clarified not-a-mod error when the SMAPI installer is in the `Mods` folder.
+
+* For modders:
+  * Added `PerScreen<T>` utility and new `Context` fields to simplify split-screen support in mods.
+  * Added screen ID to log when playing in split-screen mode.
+
+* For the Console Commands mod:
+  * Added `furniture` option to `world_clear`.
 
 ## 3.7.6
 Released 21 November 2020 for Stardew Valley 1.4.1 or later.